Moved StorageOptions into internal
