+++ log.go
@@ -21,6 +21,7 @@ import (
 	"time"
 
 	f_log "github.com/transparency-dev/formats/log"
+	"github.com/transparency-dev/trillian-tessera/internal/options"
 	"golang.org/x/mod/sumdb/note"
 	"k8s.io/klog/v2"
 )
@@ -32,33 +33,11 @@ import (
 // in an appropriate manner (e.g. for HTTP services, return a 503 with a Retry-After header).
 var ErrPushback = errors.New("too many unintegrated entries")
 
-// NewCPFunc is the signature of a function which knows how to format and sign checkpoints.
-type NewCPFunc func(size uint64, hash []byte) ([]byte, error)
-
-// ParseCPFunc is the signature of a function which knows how to verify and parse checkpoints.
-type ParseCPFunc func(raw []byte) (*f_log.Checkpoint, error)
-
-// EntriesPathFunc is the signature of a function which knows how to format entry bundle paths.
-type EntriesPathFunc func(n, logSize uint64) string
-
 // IndexFuture is the signature of a function which can return an assigned index or error.
 //
 // Implementations of this func are likely to be "futures", or a promise to return this data at
 // some point in the future, and as such will block when called if the data isn't yet available.
 type IndexFuture func() (uint64, error)
-
-// StorageOptions holds optional settings for all storage implementations.
-type StorageOptions struct {
-	NewCP   NewCPFunc
-	ParseCP ParseCPFunc
-
-	BatchMaxAge  time.Duration
-	BatchMaxSize uint
-
-	PushbackMaxOutstanding uint
-
-	EntriesPath EntriesPathFunc
-}
 
 // WithCheckpointSignerVerifier is an option for setting the note signer and verifier to use when creating and parsing checkpoints.
 //
@@ -76,7 +55,7 @@ type StorageOptions struct {
 // as the checkpoint Origin line.
 //
 // Checkpoints signed by these signer(s) and verified by the provided verifier will be standard checkpoints as defined by https://c2sp.org/tlog-checkpoint.
-func WithCheckpointSignerVerifier(s note.Signer, v note.Verifier, additionalSigners ...note.Signer) func(*StorageOptions) {
+func WithCheckpointSignerVerifier(s note.Signer, v note.Verifier, additionalSigners ...note.Signer) func(*options.StorageOptions) {
 	origin := s.Name()
 	for _, signer := range additionalSigners {
 		if origin != signer.Name() {
@@ -84,7 +63,7 @@ func WithCheckpointSignerVerifier(s note.Signer, v note.Verifier, additionalSign
 		}
 
 	}
-	return func(o *StorageOptions) {
+	return func(o *options.StorageOptions) {
 		o.NewCP = func(size uint64, hash []byte) ([]byte, error) {
 			// If we're signing a zero-sized tree, the tlog-checkpoint spec says (via RFC6962) that
 			// the root must be SHA256 of the empty string, so we'll enforce that here:
@@ -116,8 +95,8 @@ func WithCheckpointSignerVerifier(s note.Signer, v note.Verifier, additionalSign
 }
 
 // WithBatching enables batching of write requests.
-func WithBatching(maxSize uint, maxAge time.Duration) func(*StorageOptions) {
-	return func(o *StorageOptions) {
+func WithBatching(maxSize uint, maxAge time.Duration) func(*options.StorageOptions) {
+	return func(o *options.StorageOptions) {
 		o.BatchMaxAge = maxAge
 		o.BatchMaxSize = maxSize
 	}
@@ -127,8 +106,8 @@ func WithBatching(maxSize uint, maxAge time.Duration) func(*StorageOptions) {
 //
 // maxOutstanding is the number of "in-flight" add requests - i.e. the number of entries with sequence numbers
 // assigned, but which are not yet integrated into the log.
-func WithPushback(maxOutstanding uint) func(*StorageOptions) {
-	return func(o *StorageOptions) {
+func WithPushback(maxOutstanding uint) func(*options.StorageOptions) {
+	return func(o *options.StorageOptions) {
 		o.PushbackMaxOutstanding = maxOutstanding
 	}
 }
